refactor(runtime): isolate native library entrypoint emission
