+++ CMakeLists.txt
@@ -515,6 +515,7 @@ set(COPPERFIN_RUNTIME_PIPELINE_SOURCES
     src/runtime/runtime_pipeline_library_export_request_serialization.cpp
     src/runtime/runtime_pipeline_library_export_process_launch.cpp
     src/runtime/runtime_pipeline_library_export_response_handling.cpp
+    src/runtime/runtime_pipeline_library_export_entrypoints.cpp
     src/runtime/runtime_pipeline_library_export_manifest.cpp
     src/runtime/runtime_pipeline_native_wrapper_process.cpp
     src/runtime/runtime_pipeline_fxp_and_archive_manifest.cpp
