+++ src/runtime/runtime_pipeline_support.h
@@ -207,6 +207,7 @@ void append_native_wrapper_bridge_model_source(std::ostringstream& stream);
 void append_native_wrapper_request_serialization_source(std::ostringstream& stream);
 void append_native_wrapper_process_launch_source(std::ostringstream& stream);
 void append_native_wrapper_response_handling_source(std::ostringstream& stream);
+void append_native_wrapper_library_entrypoint_source(std::ostringstream& stream, const RuntimePackagePlan& plan);
 std::string build_module_definition_source(const RuntimePackagePlan& plan);
 std::string build_native_wrapper_source(const RuntimePackagePlan& plan);
 std::string build_native_wrapper_cmake_source(const RuntimePackagePlan& plan);
